Sakila_Flask_Database_Assignment/requirements.txt
mysql_connector_python
flask
pandas
os